+++ src/main/java/com/geekutil/modules/sys/mapper/PermissionMapper.java
@@ -16,7 +16,7 @@ import java.util.List;
  */
 public interface PermissionMapper extends BaseMapper<Permission> {
 
-    List<Permission> getListByRoleIds(@Param("roleList") List<Integer> roleList);
+    List<Permission> getListByRoleIds(@Param("roleList") List<Long> roleList);
 
     List<Permission> getListByRoleId(@Param("id") Long id);
 }
